+++ NEWS.md
@@ -14,6 +14,7 @@ v0.x.x - 202x-xx-xx
 - Add the `-F`/`--format` option. Supports `default` and `gpx`.
 - `--selftest` and `--valgrind` now accept an optional argument to 
   control which test categories to run: `exec`, `func`, or `all`.
+- Fix bug when `dist` printed "-nan" for certain antipodal points.
 - Add `TESTS` variable for use with `make`. This variable is used as 
   `--selftest`/`--valgrind` argument in all relevant `make` commands.
 - Add `make tlokall`, prints the result of `make tlok` with and without